Add files via upload
Mind & Body is an interactive fitness appointment booking website that seamlessly blends aesthetics and functionality. Using HTML, CSS, and JavaScript, users can easily explore fitness services, schedule appointments with certified instructors, and enjoy a hassle-free experience. With the integration of Koalendar, real-time availability and booking are made effortless. Immerse yourself in a soothing design, access detailed program information, and embark on a journey to a healthier mind and body with Mind & Body.